+++ src/output/shallow/option.rs
@@ -2,19 +2,25 @@ use core::borrow::Borrow;
 
 use crate::output::*;
 
-type Mix<T> = Shallow<Option<&'static T>>;
+type Ref<T> = Shallow<Option<&'static T>>;
+type Mut<T> = Shallow<Option<&'static mut T>>;
 
-type Response<T> = Option<Box<dyn Borrow<T> + Send + Sync>>;
+type RefResponse<T> = Option<Box<dyn Borrow<T> + Send + Sync>>;
+type MutResponse<T> = Option<Mutable<Box<dyn Borrow<T> + Send + Sync>>>;
 
-impl<T: ?Sized> Kind for Mix<T> {
-    type Return = Response<T>;
+impl<T: ?Sized> Kind for Ref<T> {
+    type Return = RefResponse<T>;
 }
 
-impl<T: ?Sized> Return for Mix<T> {
+impl<T: ?Sized> Kind for Mut<T> {
+    type Return = MutResponse<T>;
+}
+
+impl<T: ?Sized> Return for Ref<T> {
     type Type = Option<Box<dyn Borrow<T> + Send + Sync>>;
 }
 
-impl<T: ?Sized + 'static> GetOutput for Response<T> {
+impl<T: ?Sized + 'static> GetOutput for RefResponse<T> {
     type Output<'u> = Option<&'u T>
         where
             Self: 'u;
@@ -24,13 +30,23 @@ impl<T: ?Sized + 'static> GetOutput for Response<T> {
     }
 }
 
+impl<T: ?Sized + 'static> GetOutput for MutResponse<T> {
+    type Output<'u> = Option<&'u mut T>
+        where
+            Self: 'u;
+
+    fn output(&self) -> Option<Self::Output<'_>> {
+        None
+    }
+}
+
 macro_rules! into {
     ($trait:ident, $method:ident) => {
-        impl<T0, T: ?Sized + 'static> $trait<Mix<T>> for Option<T0>
+        impl<T0, T: ?Sized + 'static> $trait<Ref<T>> for Option<T0>
         where
             T0: Borrow<T> + Send + Sync + 'static,
         {
-            fn $method(self) -> OutputResult<Response<T>> {
+            fn $method(self) -> OutputResult<RefResponse<T>> {
                 Ok(match self {
                     Some(val) => Some(Box::new(val)),
                     None => None,