+++ src/value_chain.rs
@@ -8,7 +8,7 @@ use crate::alloc::Box;
 /// This allows retaining shared references to inserted values.
 #[derive(Default)]
 pub struct ValueChain {
-    root_node: OnceCell<Node>,
+    root: OnceCell<Node>,
 }
 
 impl ValueChain {
@@ -18,8 +18,21 @@ impl ValueChain {
         node.value.as_ref().downcast_ref::<T>().unwrap()
     }
 
+    pub fn add_mut<T: Any + Send + Sync>(&mut self, value: T) -> &mut T {
+        // note: There is no need for keeping the old chain.
+        // All those references are out of scope when add_mut is called.
+        self.root = Node::new(value).into();
+
+        self.root
+            .get_mut()
+            .unwrap()
+            .value
+            .downcast_mut::<T>()
+            .unwrap()
+    }
+
     fn push_node(&self, mut new_node: Node) -> &Node {
-        let mut cell = &self.root_node;
+        let mut cell = &self.root;
         loop {
             match cell.try_insert(new_node) {
                 Ok(new_node) => {
@@ -36,7 +49,7 @@ impl ValueChain {
 
 impl Drop for ValueChain {
     fn drop(&mut self) {
-        if let Some(node) = self.root_node.take() {
+        if let Some(node) = self.root.take() {
             drop(node.value);
             let mut cell = node.next;
 
@@ -72,4 +85,11 @@ fn it_works() {
     assert_eq!(&1, first);
     assert_eq!(&"", second);
     assert_eq!(&42.0, third);
+}
+
+#[test]
+fn it_works_mut() {
+    let mut value_chain = ValueChain::default();
+    let first = value_chain.add_mut(1);
+    *first += 1;
 }