+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Basic support for mutable outputs ([#49](https://github.com/audunhalland/unimock/pull/49))
 
 ## [0.6.2] - 2024-03-27
 ### Fixed